Final refactoring of plist : now each tag is a class. Rename strerror to efiStrError because of conflict when run cpp_tests on Mac. Switch base64_decode_block to long to avoid warning and cast. Correct GetTableType4(). Size was used without being initialized.

+++ rEFIt_UEFI/Platform/b64cdecode.h
@@ -48,7 +48,7 @@ int base64_decode_value(char value_in);
 
 
 
-int base64_decode_block(const char* code_in, const int length_in, char* plaintext_out, base64_decodestate* state_in);
+long base64_decode_block(const char* code_in, const int length_in, char* plaintext_out, base64_decodestate* state_in);
 
 
 UINT8
